requirements.txt
pysqlite3
langchain_huggingface
langchain_chroma
langchain_groq
streamlit
python-dotenv
boto3
sentence-transformers
requests